requirements.txt
langchain>=0.2.0
langchain-groq
langchain-community>=0.0.30
pydantic
gradio
groq
python-dotenv
